refactor: Add 'request_join' method to AppCall
This commit adds a new method called 'request_join' to the AppCall struct in the 'app.rs' file. This method allows the user to request to join the canister. It takes a name parameter and returns an OperationId. The purpose of this change is to enhance the functionality of the AppCall struct by providing a way to request to join a canister.

Co-authored-by: [Author Name]